requirements.txt
pandas
selenium
webdriver-manager
openpyxl
python-dotenv